Fixed Redux example
